Run jobs in separate processes
This is meant to isolate all the nonsense applgrid does and load a PDF
and a grid for each computation of the convolution (even if they are
implemented as singletons (!)). It is to be tested if this works or not.

Datasets will be generated in the same
format, but will be ran in parallel.