First steps for implementing installgrids
It is more difficult because breaks the checks done by parse_pdfsets. To
fix it, the grid names should be computed at check time and available
for the next actiongroup. Doing this properly requires a non trivial
rewrite as well as breaking the independence between actiongroups
(althoughr prefixes kind of do this).

Possibly Config shuold be split in ConfigParser and Config and the
parser methods should be nermal instance methods rather than
@classmethods.

Conflicts:
	src/smpdflib/actions.py