More or less working iterative selection
trynewthing.py shows the new idea. Interface has to be polished and
integratet properly with the rest of the code. An automatic selection
criteria is still missing.More or less working iterative selection

trynewthing.py shows the new idea. Interface has to be polished and
integratet properly with the rest of the code. An automatic selection
criteria is still missing.
reº