Upgraded version of the applwrap interface
Now it is compatible with Python 3 only.

The core.PDF object is responsible for getting the xqf values from
applwrap and LHAPDF is not involved. Exactly one PDF object is loaded at
once per process.

Also several functions were added to the applwrapper, namely to retrieve
xfq, min Q and de installed paths.

The functionality of the lhagrids module is this integrated in core.