Improve app initialization
The main loop is now in the ations module, ehere it should belong (but
there is a refactoring to do there as well). The initalization of the
app (multiprocessing, mpl backend and styles) is now in its dedicatd
module, so it is easy to use for another script, as exemplified by
smallthing.py.