Implement recomputation of errors
First we return a structure from get_smpdf_lincomb, as it gets too
complicated.

We define a TooMuchPrecision error (the only runtime error allowed so
far) which will be raised if the SMPDF cannot be computed with the
required precision.

After the real results are computed, the new target tolerances are set
according to:

newtol = smpdf_tolerance - (real_error[bad_bins] -
                            np.array(obs_errors)[bad_bins])

That is, the difference between linear and real is assumed to stay
constant, and we substract it from the tolerance.