Add support for PDF plots
Support base_pdf as the other tools, hatches and transparencies (but
there is a bug in matplotlib atm) and smarter automatic rescaling.

PDF values are converted to results, so errors can be computed with
the existing interface.

A PDFValues "Observable" was also added, but its value is unclear
(apart from allowing result.meanQ wich is not that used anyway).

Accompanying utilities are also present, e.g. to iter hatches and
break grid values by flavour.