Improve the save_figure interface
In general this is a little bet of a mess. save_figures was for some
reason responsible for executing the generators, which can be done
much more easily in the respective actions.

In any case it should be generalized in such a way that the actions
become needless (as they do pretty much nothing the moment).