Implement pdfplots action
Requires the energy Q and allows to select a subset of flavors.

Maybe we should think of more options like selecting xranges (logspace
between extrema?).

Contrary to the convention in the rest of SMPDF we generate a big
figure rather than many by flavor, but it should be fairly easy to
change that.