Fix meanQ attribute for labelled bins
When binlabels is not set by a range (such as for prediction
observables), the meanQ is not necessarily indexed by the binlabel.
The best way is to iterate the Qs using `iter` so that how is it
indexed becomes irrelevant (probably would be better to make it
a Series with the same Index rather than a list).