Use APFEL convolution
This is yet not general enough to be merged, and a few APIs would be
needed to select the settings and so on.

Yet, it is hopefully correct. The initialization and PDF selection is
done in each process that runs convolute, and we error out if for
some reason it doesn't start a new shell.